requirements.txt
gmusicapi
oauth2client
flask
google-cloud-datastore